backend/requirements.txt
# Python Dependencies for Flask LMS Cloud Backend

# Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2

# Database
cx-Oracle==8.3.0
SQLAlchemy==2.0.20

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# Environment Management
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# API Documentation
flask-swagger-ui==4.11.1

# Rate Limiting
Flask-Limiter==3.5.0

# File Upload
python-magic==0.4.27

# Email (Optional)
Flask-Mail==0.9.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
pylint==2.17.5

# Production Server
gunicorn==21.2.0

# Monitoring
prometheus-flask-exporter==0.22.4

# Utilities
requests==2.31.0
